backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Faces BE"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
django-extensions = "^3.2.3"
django-redis = "^5.4.0"
djangorestframework = "^3.15.1"
redis = "^5.2.0"
django-environ = "^0.11.2"
pillow = "^11.0.0"
django-rq = "^3.0.0"
dlib = "^19.24.6"
face-recognition = "^1.3.0"
setuptools = "^75.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
